22 select an image pane and assign image from current row keys ctrl+19

+++ src/data_table.rs
@@ -24,7 +24,7 @@ impl DataTable {
                     .map(|cell| Cell::from(cell.clone()))
                     .collect::<Vec<_>>();
                 let mut table_row = Row::new(row_cells);
-                if index == app.current_row_index as usize {
+                if index == app.current_row_index {
                     table_row = table_row.reversed();
                 }
                 table_row
